Simplify string escaping in DOTBuilder
The code replaces \" with \ "
and " with \"
and \n with something else.
Due to the unnecessary use of regexps, lots of quoting was necessary.

The first replacement was added in r1251 (ba9549b4),
the third replacement was added in r40,
and the second replacement was already present in r1
(it is from 2007 and older than CPAchecker's version history).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@37857 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c