Add a first version of the TraceAbstractionCPA.
The approach is based on 'Software Model Checking for People Who Love Automata'
[Heizmann, Hoenicke, Podelski] (DOI: 10.1007/978-3-642-39799-8_2), and is
essentially about automata that are built from interpolants and
afterwards used for successive CEGAR iterations.

This commit is mostly about storing the intermediate state as well as to give
a first idea of the direction the implementation is heading. Some pivotal steps
are therefore still missing, like the proper handling of predicates.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@37882 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c