Switch from Guava's Optional to the Java Optional in the CFA package
Having one Optional type is much nicer than mixing two
(for example, some code had to manually use the fully-qualified class
name for this).
The Java Optional is the recommended one.
Guava's Optional has the advantage that it is serializable,
but we do not need that here.
For the few cases where it is relevant, we switch to @Nullable fields.

Additionally, we change the return type of CFAEdge.getRawAST()
from Optional<? extends AAstNode> to Optional<AAstNode>.
The former allowed us to use a more specific type
(like Optional<CExpression>) in subclasses of CFAEdge,
but getRawAST() is intended only for cases
where the exact type does not matter any.
And since the previous two commits,
only such users remain for this method.

Thanks to Valentin Port for bringing both issues up
and suggesting to fix them!

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@37924 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c