Remove redundant values from PredicateOrderingStrategy
DISABLE made it do nothing, i.e., variables were in creation order,
and CHRONOLOGICAL made it call RegionManager.reorder()
with a parameter that tells it to put variables in creation order
(i.e., do nothing).
This is confusing and should be simplified.
Now only CHRONOLOGICAL is left and continues to be the default.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@37934 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c