Attempt to fix one of the problems with ThreadSafeTimerContainer
If I understand everything correctly, adding this one call adds a
"happens-before" relationship between that should ensure that we see
timer actions like stop() properly from the thread that aggregates the
statistics.

Note that this heavily depends on the internals of the Timer class,
which does not guarantee the current behavior!
But at least for now it makes ThreadSafeTimerContainer less problematic.

Of course, this does not resolve the other races like timers being
started/stopped while measurements are read, it adds only memory
visibility.

A full fix would require either fully synchronizing TimerWrapper
(probably too slow for short measurements in hotspot loops)
or some other solution that would still have to be found.
Probably the code should also not depend on Timer but do everything on
its own.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@37954 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c