Refactor CPABuilder: introduce abstraction for configuration
So far, we parsed configuration options and immediately created CPA
instances for each of them.
Now we first collect all configuration options that declare the tree
of CPAs and store them in a datastructure.
After this is finished we create the actual CPA instances.

The main goal of this is a preparation for the next commit,
but it also has the advantage of making the code easier to understand
and avoids mixing handling configuration options,
sanity checks on the configuration,
and instance creation.

In order to make this commit as easy as possible to understand,
the least possible amount of changes was implemented for now,
even though some refactorings and renamings would further improve the code.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/cpabuilder-specification@37987 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c