Change how and where CPABuilder inserts CPAs for specification automata
Before r30967 / db4d4d4a, the following behavior was used:
Iterate in a depth-first pre-order manner over the configured tree of CPAs
and pick the first CPA that either already has the option ".cpas" set
or has no configured children and the alias "CompositeCPA"
(we refer to this CPA as the "insertion point" in the following).
The CPAs for specification automata are added as children of the
insertion point, and if the insertion point has no other children,
a LocationCPA is added as well.

The above mentioned commit changed this and added the following
behavior:
If the configured tree of CPAs contains a CPA instance with the alias
"ControlAutomatonCPA" and if this instance comes before the insertion point
in the depth-first pre-order iteration, then replace it with the first
of the CPAs for specification automata
(completely ignoring the configured class and other options
for the CPA instance with alias "ControlAutomatonCPA").
The other CPAs for specification automata would be added
at the usual insertion point.

This behavior was undocumented and dangerous for users
(because configured CPA instances were silently replaced with something
else if their alias happened to be "ControlAutomatonCPA").
Thus we remove this behavior again,
even though it requires changes to configurations using the DCACPA
(for LTL model checking), which relied on this behavior.
More discussion about this is at
https://gitlab.com/sosy-lab/software/cpachecker/-/commit/20eefbfc75d356b86219d034827f59bc26b87bf9#note_648305199

As replacement, we now provide a (documented) way to specify
where CPAchecker should insert CPAs for specification automata:
the placeholder "$specification".
Note that if there is a CPA instance with the alias "$specification",
this will not be interpreted as the placeholder and the CPA instance
will be configured as usual
(this fixes the dangerous current behavior).
Furthermore, we no longer provide the counterintuitive, error-prone,
and unnecessary possibility to put one of the CPAs for specification
automata in a different place than the others.
All CPAs for specification automata will be added where "$specification"
is used, and if there are more than one such CPAs but the place of
"$specification" accepts only one CPA, this will be an error.
The placeholder also must not occur more than once.

If no placeholder is given, the old standard behavior from before r30967
will be used.
Care was taken to replicate the behavior closely,
even though with the new code it would be easy to use a different
more intuitive behavior such as adding the CPAs for specification
automata to the top-most composite CPA.

The configuration file ltl.properties is also adjusted as necessary
in this commit.
We remove its (wrong and redundant) default specification
such that when executing it without an LTL specification
the error message will correctly indicate the missing specification
(which is possible now).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/cpabuilder-specification@37988 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c