Remove LogManager from VariableClassification and GlobalInfo
GlobalInfo is a workaround for serialization of certain data structures
that cannot be serialized without access to some central state.
It has several problems (e.g., it will immediately fail in case of a
more complex analysis like a portfolio) and we should use it as rarely
as possible.
To do so, we can avoid storing non-data classes (like LogManager)
in fields of data classes (like VariableClassification).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@38003 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c