Remove option analysis.algorithm.CBMC, i.e., executing CBMC as analysis
This option allowed to just call CBMC from CPAchecker
instead of using one of the analyses implemented in CPAchecker itself.

The main use case was using CBMC as part of a sequential combination of
analyses as done in our technical report in 2011:
https://arxiv.org/abs/1109.6926
This use case was broken by revision 20824 / 5a444c91 five years ago
(creating ExternalCBMCAlgorithm from CoreComponentsFactory never worked
because it would fail when creating CPAs due to no CPAs being
configured).

It was also possible to use CBMC from CPAchecker outside of a sequential
combination, which would make CPAchecker a trivial wrapper around CBMC.
The config file CBMC.properties for this was marked as unmaintained
three years ago.
Furthermore, this use case is pretty irrelevant,
because one could just execute CBMC directly without CPAchecker.

The use of CBMC as analysis from within CPAchecker always had problems.
For example, most configuration options given to CPAchecker were
ignored, in particular the specification
(ExternalCBMCAlgorithm had its own configuration option for an error
label instead).
This means that users who use "-CBMC -spec ..." would get an analysis
that did not verify what it should.

The fact that ExternalCBMCAlgorithm exists also always created the need
for special cases in several circumstances due to it being the only
configuration where not even a CFA was created.
This also caused crashes, e.g., in ReportGenerator.

So it is better to remove this.
Note that this is NOT related to executing CBMC as a counterexample
checker.
There is a separate option for that and this always worked better than
executing CBMC as main analysis and is much more useful,
so we keep it.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@38004 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c