Fix adding specifications for configs with two or more CompositeCPAs
This fixes a bug from r37987 / 8e75ac3b.
The previous code added the CPAs for specification automata
to the first composite CPA that it created,
which was the left-most and deepest composite CPA in the tree of CPAs
(because it created leaf CPAs before higher levels).
I had misinterpreted this to be depth-first pre-order,
but of course this is depth-first post-order.
So of course we also need post-order in the new code
(where it shows how nice the refactoring from that commit is
because now this fix is a small and almost declarative change).

Affected by the bug were only configurations with two or more composite
CPAs, probably only our slicing configs in practice.

Should have been part of !58.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@38022 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c