Improve error message in TraceAbstractionCPA
- Error messages should typically be a single line. Otherwise they get
split when being logged and parts will often be lost when the log is
analyzed by a tool, for example.
- Providing a suggested configuration fix is a great idea, but not easy,
the suggested fix was not always correct because aliases of CPAs in the
config file can be any string. Furthermore, adding a line is not
helpful, because the respective line has to exist already. After all, a
child CPA was configured, it is just of the wrong type.
- Giving the information about the current wrong state is often useful.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@38050 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c