Add AbstractSerializableSingleWrapperState
Removing Serializable from AbstractSingleWrapperState was a good idea
because we should not burden all subclasses with the hassle of
serialization, even if they don't need it,
but it turns out that serializable subclasses cannot make use of
AbstractSingleWrapperState as it is.
The easiest way to work around this is to copy the class
(which is luckily not that large).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@38067 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c