Propagate InterruptedException in AbstractionFormula.asExpressionTree
This also requires adding "throws InterruptedException"
in a few more places, but nowhere problematic.

We finally catch the exception in the statistics code
where output files are written,
because if such an exception happens during statistics
and writing output files we want to continue with the statistics and
output files from other components.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@38073 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c