Rename two methods of MemoryLocation
MemoryLocation invents a specific string format
for when it should be dumped into a single string,
and callers must be aware about this,
so the method names have to reflect this.

As can be seen when looking at the diff of this commit,
many callers actually pass string arguments to valueOf()
that are not of this format,
but just happen to work nevertheless.
Where possible, I will clean this up.
For other cases we can hope that it will be cleaned up later
and that at least due to the new method name
no more code will be added where this method is misused.

NOTE: This commit is just a renaming!
If you wonder why certain code is changed by this commit
and now calls a weird method, this behavior did already exist,
it was just undocumented.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@38080 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c