Refactor main constructor of MemoryLocation
We need a constructor that allows creating MemoryLocations for global
variables without attempting to parse the given string as a qualified
name. Thus the constructor has to allow passing null as function name.

All callers of this constructor can easily be adapted to have null
checks where needed. This also allows easier copying of instances
and in one cases avoids that passed strings will be parsed twice.

Furthermore, the field functionName was nullable, so mark it as such.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@38082 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c