Remove PointerToMemoryLocation
This class was previously useful, but is now an empty subclass of
MemoryLocation and can be removed according to its author.

While PointerToMemoryLocation.valueOf() parsed the given identifier
like it was a qualified name, the only caller actually just passes a
field name, so MemoryLocation.forIdentifier() is equivalent,
no need to attempt to parse it because field names never contain "::".

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@38084 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c