Rename remaining MemoryLocation.valueOf() overloads
These methods took a qualified name and parse it,
so the method name should reflect that.

In order to not confuse future developers of this class,
the parsing should also be done in a static method
and not as a constructor overload.
Otherwise it is too easy to call that constructor
when attempting to create a MemoryLocation for a local variable.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@38088 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c