Add another utility factory method to MemoryLocation
Often we have variable declarations and want a MemoryLocation for them,
so we can provide the appropriate method.
Note that the callers changed by this commit all called a different
factory method of MemoryLocation that would actually parse the given
string differently than qualified names should be parsed.
This confusion was due to the many factory names with the same name
and worked because qualified names were a subset of the valid values
that could be passed to this factory method.
But now we handle this properly.

The callers have been changed using Refaster.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@38090 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c