Remove useless synchronization from AggregatedReachedSets
The field reachedSets was used for synchronized only once.
In AggregatedReachedSets there is no need for synchronization here,
and the subclass AggregatedThreadedReachedSets adds its own lock anyway.

synchronized for AggregatedReachedSetManager was used only in a single
place, and there also the additional lock was used.

The ConcurrentHashMap that was used for reachedSets in
AggregatedReachedSetManager was never used without the explicit lock
held, so there is no need for a concurrent data structure.
Instead, we can benefit from using one with deterministic iteration
order.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@38115 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c