Refactor AggregatedReachedSets and make it abstract
It was too confusing that this class was used as both concrete instance
and as superclass, and that it stored a set that was sometimes immutable
(when the class itself was instantiated) and sometimes mutable (when
used as superclass) and that all modifications to that set where only
done in the subclass.

So now the class is abstract and has no state,
and each subclass stores and manages its state as appropriate.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@38117 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c