Provide a CPA instance when creating reached sets
As discussed in
https://gitlab.com/sosy-lab/software/cpachecker/-/commit/21fb48966e20471c211e2a636fa0b26a78d6e063#note_652539366
our "main" CPA instance (the one used by the CPAchecker class)
is not always the CPA instance that corresponds to the abstract states
in the resulting reached set from a CPAchecker run,
for example not in cases of composite analyses and uses of
ForwardingReachedSet.
This could lead to problems if code analyzes the reached set
and applies the wrong CPA to it.

The best fix seems to store a reference to the corresponding CPA
in ReachedSet, and this commit prepares this by adding new factory
methods for ReachedSet that take a CPA as parameter.
Most calls of the existing factory methods can be changed easily to also
pass a CPA instance, and this commit also does that.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/store-cpa-in-reachedset@38125 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c