Detect more kinds of missing results in WebClient mode of benchmark.py
Sometimes the VerifierCloud returns results without a log, for example.
For such incomplete results we also need to mark them using set_error()
such that they won't get interpreted as complete results by some tool.

Furthermore, we use this to still write at least a dummy log with the
command line for such runs, because otherwise the command line would not
be accessible to the user (and BenchExec would crash without a log).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@38133 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c