Revert "Simple Waitlist-classes implement Serializable now."
This reverts commit r27657 / 636e245ce88ba5d81f8d7fa4bd78ea5897a65524.

Serializing waitlists is never used in CPAchecker,
and because Java's serialization is so complex and often problematic,
it is better to make only those classes serializable that actually need
to be in order to reduce the maintenance effort.

For example, this reduces the dependency on CFAInfo,
which exists primarily for serializing LocationStates.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@38161 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c