Revert "Some ReachedSet-classes implement Serializable now."
This reverts commit r27656 / 19a8073ff4d3104260d005ce7934a1ed22fe3848.

Serializing reached sets is never used in CPAchecker,
and because Java's serialization is so complex and often problematic,
it is better to make only those classes serializable that actually need
to be in order to reduce the maintenance effort.

For example, this allows us to make all fields final
and reduces the hassle for subclasses.
It also simplifies !59.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@38162 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c