Revert "VariableTrackingPrecision is serializable now."
This reverts commit r27564 / fa86295857ac0150b08c4cb384a4263032889a1e.

Serializing precisions is never used in CPAchecker,
and because Java's serialization is so complex and often problematic,
it is better to make only those classes serializable that actually need
to be in order to reduce the maintenance effort.

For example, this allows us to make more fields final and simplify the
CFAInfo class.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@38164 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c