Revert "Serialization for SingletonPrecision."
This reverts commit r27540 / 8cf6eb9cc47a4073b00c54eab0fd51cab13acc22.

Serializing precisions is never used in CPAchecker,
and because Java's serialization is so complex and often problematic,
it is better to make only those classes serializable that actually need
to be in order to reduce the maintenance effort.

For example, this class could be made serializable more easily
by converting it into an enum.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@38165 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c