Store CPA in ReachedSet
The previous commits and !59 explain why this is necessary.

We do not provide access to the CPA instance from UnmodifiableReachedSet
because this is currently not necessary and more effort.
We just need to expose ReachedSet instead of UnmodifiableReachedSet
from CPAcheckerResult, but this is ok because whoever has access to
CPAcheckerResult owns it and the contained data.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/store-cpa-in-reachedset@38167 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c