Avoid using GlobalInfo in ARGBasedPartialReachedSetConstructionAlgorithm
The CPA instance that is stored there is not necessarily the correct one
in case a meta analysis like RestartAlgorithm or ParallelAlgorithm is
used, cf.
https://gitlab.com/sosy-lab/software/cpachecker/-/commit/21fb48966e20471c211e2a636fa0b26a78d6e063#note_652502045

So now we pass the CPA instance from outside. This is part of #909.

In the end, there are two places where we need to provide a CPA
instance:
- ProofCheckAlgorithm: Here a CPA instance already exists and is passed
  to PCCStrategyBuilder.buildStrategy(), we can use the same.
- ProofGenerator: Here we can use ReachedSet.getCPA() from !59.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/store-cpa-in-reachedset@38169 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c