Use ReachedSet.getCPA() in ResultCheckAlgorithm
In most cases, this is the same CPA instance as passed to
ResultCheckAlgorithm by CoreComponentsFactory.
However, in meta analyses that use ForwardingReachedSet,
it could differ, and then the CPA from the reached set would be the
correct one to use.
This is also more consistent with what ProofGenerator does
since the previous commit.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/store-cpa-in-reachedset@38170 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c