Improve how property files are handled in CPAchecker
This is basically a continuation of the work in r24988 (6f8a7ed0).
When we first started supporting property files,
we allowed them only as command-line arguments and mapped them to regular
specification files (with automata) as soon as possible.
At some point we needed information about which property was given to
CPAchecker (for writing witnesses), so this information was stored
additionally.
Then (in the commit mentioned above) we moved the parsing of property
files (and their mapping to specification files) to a later point,
such that property files were not only accepted on the command line.

However, property files were still mapped to specification files by the
CPAMain class. This had several disadvantages:
- The automata files resulting from property files and the properties
  themselves both needed to be passed to the CPAchecker class,
  and this was done in two different ways (as configuration option and
  as parameter to CPAchecker.run()).
- When writing a witness, we needed to distinguish between those
  automata that were directly given to CPAchecker as specification files
  and those automata that resulted from a specification file that was
  added due to a given property file.
- Our portfolio algorithm needs to pass the specification to its
  subanalyses, but it could do so only by passing the list of
  specification files, which would loose all information about property
  files. This means that witnesses from subanalyses of the portfolio
  algorithm would not contain the expected property definition.
- When using the CPAchecker class directly (i.e., not from command line)
  property files were entirely unsupported.

Now the property files are handled like regular specification files,
and all these problems are fixed.
The configuration option "specification", which is read by the
CPAchecker class, can now contain both kinds of files,
and the class Specification will correctly parse all of them.
The information stored in the Specification instance will correctly
reflect this, i.e., its method getSpecFiles() will now also return
property files if those were given.
The fact that property files are mapped to specification files in order
to get an automaton for them is no longer visible anywhere outside the
parsing method of the Specification class.

Except for the fixed problem with the portfolio algorithm,
there should be no noticeable difference for users.

Even after this commit,
the CPAMain class still parses property files because we also want to
switch to a different configuration file depending on the property.
So this feature is still supported only on the command line,
but it might possible to move this to the CPAchecker class in the future.
The fact that property files are now parsed twice is not problematic
because these files are typically small and in CPAMain, the LTL part of
these files is not parsed, so it is fast.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/integrate-property-files-in-specification@38173 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c