Fix bug in code for writing output files
If CPAchecker's result is NOT_YET_STARTED (because of a parsing failure)
we do not print statistics and do not write output files.
However, for the latter the check was missing.
This was not noticed before because CPAchecker actually always
returned a reached set, although it was not intended to do so.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@38179 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c