Rename interfaces.Property to interfaces.Targetable.TargetInformation
We have two central interfaces named "Property",
so we need to rename at least one of them.
Furthermore, it is confusing that Targetable.getViolatedProperties()
does not actually return anything that is related to the properties
of our Specification instance, but in the end just human-readable
descriptions.
While it might be useful to add a connection from
getViolatedProperties() to the actual Property instances,
we have no need for this now.

The interface Targetable is currently also inconsistent,
because in general it is about the term "target" and has the method
isTarget(), but then there is the one method getViolatedProperties()
that does not use the term "target" but something else.

So we now rename interfaces.Property to TargetInformation
and Targetable.getViolatedProperties() to getTargetInformation().
Summary of advantages:
- Targetable is now more consistent.
- Avoids two common interfaces with same name.
- TargetInformation is more general, for example we could easily add a
  method like "getViolatedProperty()" to it that returns a reference to
  the actual violated property instance from our specification.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@38182 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c