Nodes are relevant if they can not be merged properly.
The current way of merging nodes is most likely unsound if the node to be removed has multiple entering edges (leaving edges if merging into the sink).
An attempt has been made to make mergeNodes sound in general but at the cost of much higher time consumption of the witness export.
Instead we now refrain from merging problematic nodes and argue that they are in fact relevant because they represent locations where control flow branches or meets.
With this the shrinking of the witness automaton becomes sound and the problems mentioned in #598 and #794 no longer arise.
The witness export also gets a slight speedup at the cost of slightly bigger witnesses.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/witness-export-fixes@38328 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c