implement early propagation for BackwardAnalysis
The new Stop-Operator BlockEntryReachedStopOperator terminates a
BackwardAnalysis as soon as it reaches the block entry. In this case,
the task then
a) creates new BackwardAnalysis tasks on predecessor blocks to propagate
   the identified error condition,
b) empties the reachedSet, and re-adds only those items which have
   been in its waitlist, and
c) restarts the CPA algorithm with the modified reachedSet.
As a result, the analysis continues to explore paths which have not yet
reached block entry. As soon as they do, this procedure repeats from a).

Interrupting the analysis before it terminates by itself in order to
communciate already identified error conditions to predecessor blocks
("early propagation") becomes necessary because for blocks which contain
loops, a BackwardAnalysis does never terminate. For such blocks, the
analysis therefore remains unsound, but with early propagation it can
still identify and propagate property violations.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/block-based-task-partitioning@38465 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c