continue ForwardAnalysis after a target has been found
After a ForwardAnalysis has identified a target state and requested a
BackwardAnalysis for it, it now continues to explore the block. In order
to so, the AutomatonStates in the CompositeState are reset to their
initial states, and the algorithm gets started again. This procedure
terminates as soon as no states remain in the waitlist.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/block-based-task-partitioning@38466 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c