replace PathFormula with BooleanFormula, synchronize SSA indices
Replace all BooleanFormulas by PathFormulas to not lose the SSA maps of
each Formula. Using such maps, the new method stitchIndicesTogether
combines a block summary and an error condition to construct a formula
which remains satisfiable exactly if the context provided by the summary
does not render the error unreachable.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/block-based-task-partitioning@38472 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c