add separate task for BackwardAnalysis continuation
If a BackwardAnalysis encounters a loop, it will not terminate. In this
case, the analysis will run indefinitely and propagate new error
conditions whenever it identifies them. This situation is deliberate.

A problem which arises from this constellation is that a number of such
infinite BackwardAnalysis tasks can occupy all available threads of the
ExecutorService, thus rendering the "normal" propagation of error
conditions identified by them impossible. In order to continue such
propagation and possibly determine the program unsafe, this commit
splits BackwardAnalysis into two parts: (1) the
ordinary BackwardAnalysis and (2) a BackwardAnalysisContinuation which
starts with an already populated reachedSet and continues the analysis.
If a BackwardAnalysis reaches program entry and performs early
propagation, it now also creates a BackwardAnalysisContinuationRequest
to schedule a BackwardAnalysisContinuation which will continue the
analysis for a non-empty waitlist. This provides the scheduler with the
ability to fairly distribute execution time between infinite loop
analysis tasks and actual propagation towards program entry.

Currently, this produces severe code duplication. Should be refactored.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/block-based-task-partitioning@38474 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c