fix synchronization of SSA indices
Before this commit, stitchIndicesTogether assumed that PredicateCPA with
AnalysisDirection.BACKWARD would use decreasing indices to enumerate
assignments while traversing the program in a backward direction. This
assumption was wrong; it uses ordinary, increasing indices. As a result,
information about the value assigned to a variable closest to block
entry is represented in the formula with the highest index, not the
lwoest one.

This contribution modifies stitchIndicesTogether to honor this fact.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/block-based-task-partitioning@38503 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c