use appropriate target states during analysis
The new BlockAwareCompositeState reports itself as target state if
  1. it belongs to a ForwardAnalysis and
    1.1 its wrapped state is an error state according to the
        specification or
    1.2 its wrapped state belongs to a block exit location; or
  2. it belongs to a BackwardAnalysis and the wrapped state belongs to
     the block entry location.

This modification is a step towards applying CEGAR in both analysis,
because CEGAR refines a precision with the goal of reaching such
target states.

This also enables us to use targetStateSatCheck=true instead of
satCheck=1.

For this functionality, we have to modify PredicatePrecisionAdjustment
(again), in order to not have it compute abstractions at target states.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/block-based-task-partitioning@38512 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c