Update list of Linux bugs found with the help of CPAchecker
The page https://linuxtesting.org/results/ldv-cpachecker does not work reliably,
so we go back to listing the found bugs on our own as suggested by Evgeny.

The new list has 244 bugs that were all found by LDV with the help of CPAchecker,
and have a fix committed to the Linux kernel
(even more bugs were found but could not be analyzed yet
or where not acknowledged or fixed in Linux).
This is impressive!
Congratulations to everyone contributing to CPAchecker and LDV
who made this possible!

Big thanks to Evgeny for noticing that the previous list of bugs
was outdated and providing an uptodate list!




git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@38535 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c