add thread identifier as property to MBeam ObjectName
Each thread participating in the concurrent analysis prepares its own
PredicateCPA with an independent AbstractionManager. During the
registration of the management bean, each of these AbstractionManagers
tries to register its bean with an identical name shared aong all of
them.

Because Bean management is shared between threads, this leads to a race
condition in AbstractMBean::register().

To circumvent this problem, this contribution appends the thread name
as property within the key-properties-list of the ObjectName of the
Bean. As a result, each thread now registers an independent MBean for
the AbstractionManager.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/block-based-task-partitioning@38553 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c