Create a branch to develop single-loop transformation
There has been an implementation of this transformation in branch cfa-single-loop-transformation,
which was used to evaluate interpolation-based MC in CPAchecker.
However, the old implementation might produce a CFA that has multiple stop state in its ARG.
This new branch aims at a new implementation of the transformation to fix the problem.
The new implementation will be used in a baseline approach (SLT+LBE)
to adopt other hardware model-checking algorithms, such as interpolation sequence or IC3/PDR.



git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/cfa-single-loop-transformation-2@38574 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c