Use 'CFAEdge' instead of 'Equivalence.Wrapper<CFAEdge>' as edge type for 'MutableCfaNetwork'
CFAEdges are equal if their predecessor and successor are equal.
This would make parallel edges (i.e., edges connected to the same nodes in the same order)
impossible as they would always be equal. This was the reason for the equivalence wrapper.
However, parallel edges should never occur in CFAs, so the equivalence wrapper is unnecessary
and removed by this commit. But a CFA may contain edges that are parallel to function summary
edges. Generating these edges can be disabled via the configuration option 'analysis.summaryEdges'.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/array-abstraction-improvements@38595 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c