Rename 'ArrayAbstractionResult.Status.FAILED' to 'UNCHANGED'
It's expected that array abstraction does not work for every possible program,
so the status `FAILED` is probably not the best term for indicating that.
Status 'FAILED' is changed to 'UNCHANGED', to indicate that it is not some
unexpected failure of the array abstraction algorithm. Also, when array
abstraction is not possible, the program is not transformed (it stays unchanged).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/array-abstraction-improvements@38768 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c