Run analysis even if array abstraction is not precise
The delegate analysis should always run. If array abstraction is not possible
or the array abstraction is imprecise (and imprecision is not allowed),
the analysis should run on the original program (with non-abstracted arrays).
Previously, the analysis was only run for precise array abstractions,
so for many programs, no property was checked at all.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/array-abstraction-improvements@38769 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c