use isTarget() of BlockAwareCompositeState
In the decision whether to continue or break the analysis after
precision adjustment, the prec operator before relied on the isTarget()
method of the CompositeState, instead of the corresponding method of
BlockAwareCompositeState. As a result, target conditions specific to
the block-partitioned analysis (i.e. block entry or exit) have not been
recognized.

With this commit, it now bases the decision on the isTarget() method of
BlockAwareCompositeState.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/block-based-task-partitioning@38790 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c