Set cpa.smg.trackErrorPredicates=true where cpa.smg.trackPredicates=true
The option trackErrorPredicates was introduced in r38811 (92416492)
and is used in a case where previously trackPredicates was used.
So in order to keep the existing behavior in cases where it is not
desired to change it, we need to set the new option to the same value as
the old option in config files were the old option appears.

In r38811 this was done only for smg-ldv.properties,
now we do it also for the other config files.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@38817 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c