Mark memset and similar methods as unsupported for InvariantsCPA
This helps to avoid wrong proofs,
because these methods have side effects that are currently not handled.
We already do the same in CtoFormulaConverter of the PredicateCPA,
and InvariantsCPA is mostly used as part of a k-induction config,
so it actually should have similar behavior as CtoFormulaConverter.
Otherwise whether or not wrong proof is produced can even depend on the
timing of the k-induction and the invariant analysis that are run
concurrently.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@38844 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c