return NO_PROPERTY_CHECKED if analysis aborted
With this contibution, the Scheduler now returns NO_PROPERTY_CHECKED if
the analysis status gets queried but the analysis has not yet finished
successfully. This situation occurs if either the analysis still
runs in parallel, or it has been aborted. The status NO_PROPERTY_CHECKED
in these situations ensures that no wrong result gets reported as
overall analysis result.

Because of Jetbrains s.r.o.'s inability to implement
    https://youtrack.jetbrains.com/issue/IDEA-186988,
this commit also contains changes to the collectStatistics() method,
which should not be part of it and render this commit non-atomic and
non-compilable.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/block-based-task-partitioning@38958 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c